Refactor details page and remove complex animations for simpler layout
