+++ components/ui/animation/layout-transition.client.tsx
@@ -1,0 +1,73 @@
+"use client";
+
+import { AnimatePresence, motion, TargetAndTransition, VariantLabels } from "motion/react";
+import { useSelectedLayoutSegment } from "next/navigation";
+import { LayoutRouterContext } from "next/dist/shared/lib/app-router-context.shared-runtime";
+import { useContext, useEffect, useRef } from "react";
+
+function usePreviousValue<T>(value: T): T | undefined {
+  const prevValue = useRef<T | undefined>(undefined);
+
+  useEffect(() => {
+    prevValue.current = value;
+    return () => {
+      prevValue.current = undefined;
+    };
+  });
+
+  return prevValue.current;
+}
+
+function FrozenRouter(props: { children: React.ReactNode }) {
+  const context = useContext(LayoutRouterContext);
+  const prevContext = usePreviousValue(context) || null;
+
+  const segment = useSelectedLayoutSegment();
+  const prevSegment = usePreviousValue(segment);
+
+  const changed =
+    segment !== prevSegment &&
+    segment !== undefined &&
+    prevSegment !== undefined;
+
+  return (
+    <LayoutRouterContext.Provider value={changed ? prevContext : context}>
+      {props.children}
+    </LayoutRouterContext.Provider>
+  );
+}
+
+interface LayoutTransitionProps {
+  children: React.ReactNode;
+  className?: string;
+  initial?: boolean | TargetAndTransition | VariantLabels;
+  animate?: TargetAndTransition | VariantLabels;
+  exit?: TargetAndTransition | VariantLabels;
+  transition?: object;
+}
+
+export function LayoutTransition({
+  children,
+  className,
+  initial = { opacity: 0, y: 20 },
+  animate = { opacity: 1, y: 0 },
+  exit = { opacity: 0, y: -20 },
+  transition = { duration: 0.3, ease: [0.22, 1, 0.36, 1] },
+}: LayoutTransitionProps) {
+  const segment = useSelectedLayoutSegment();
+
+  return (
+    <AnimatePresence mode="wait" initial={false}>
+      <motion.div
+        key={segment}
+        initial={initial}
+        animate={animate}
+        exit={exit}
+        transition={transition}
+        className={className}
+      >
+        <FrozenRouter>{children}</FrozenRouter>
+      </motion.div>
+    </AnimatePresence>
+  );
+}
